requirements.txt
appdirs==1.4.3
numpy==1.12.1
olefile==0.44
opencv-python==3.2.0.7
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
six==1.10.0
